+++ tests/Contract/JsonSchema/Describer/SerializerDescriberTest.php
@@ -41,13 +41,12 @@ final class SerializerDescriberTest extends TestCase
         $describer->describeObject(new ObjectSubject(ListItemDiscriminator::class, new \ReflectionClass(ListItemDiscriminator::class)), $schema);
 
         $expectedSchema = [
-            'type' => 'object',
             'anyOf' => [
                 [
                     'description' => ListItemName::class,
                     'properties' => [
                         'type' => [
-                            'const' => 'name',
+                            'enum' => ['name'],
                         ],
                     ],
                 ],
@@ -55,7 +54,7 @@ final class SerializerDescriberTest extends TestCase
                     'description' => ListItemAge::class,
                     'properties' => [
                         'type' => [
-                            'const' => 'age',
+                            'enum' => ['age'],
                         ],
                     ],
                 ],